requirements.txt
attrs==22.2.0
bplib==0.0.6
cffi==1.15.1
coverage==7.2.0
ecdsa==0.18.0
exceptiongroup==1.1.0
fastecdsa==2.3.0
future==0.18.3
iniconfig==2.0.0
mod==0.3.0
msgpack-python==0.5.6
packaging==23.0
petlib==0.0.45
pluggy==1.0.0
pycparser==2.21
pycryptodome==3.17
pytest==7.2.1
pytest-cov==4.0.0
six==1.16.0
tomli==2.0.1
